misc: minor refactor
